+++ model/Job.php
@@ -12,14 +12,16 @@ class Job
     public $id_e;
     public $id_d;
     public $id_u;
-    /** @var int */
-    public $id_ce;
+    public int $id_ce;
     public $etat_source;
     public $etat_cible;
     public $last_message;
     public $lock;
+    public string $lock_since;
     public $id_verrou;
     public $is_lock;
+    public int $id_daemon;
+    public Daemon $daemon;
 
     public $nb_try;
     public $first_try;
@@ -27,6 +29,7 @@ class Job
     public $next_try;
 
     public $id_job;
+    public WorkerObject $worker;
 
     public function __construct()
     {
@@ -37,6 +40,7 @@ class Job
 
         $this->etat_cible = false;
         $this->id_verrou = "";
+        $this->id_daemon = 1;
         $this->next_try = date("Y-m-d H:i:s");
     }