+++ test/PHPUnit/model/DaemonSQLTest.php
@@ -1,0 +1,86 @@
+<?php
+
+declare(strict_types=1);
+
+class DaemonSQLTest extends PastellTestCase
+{
+    private DaemonSQL $daemonSQL;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->daemonSQL = $this->getObjectInstancier()->getInstance(DaemonSQL::class);
+        $this->daemonSQL->insertGlobalDaemon();
+    }
+
+    public function testCreate(): void
+    {
+        static::assertNotNull($this->daemonSQL->insertDaemon(1));
+    }
+
+    public function testGetDaemon(): void
+    {
+        $id_e = 1;
+        $id_daemon = $this->daemonSQL->insertDaemon($id_e);
+        $daemon = $this->daemonSQL->getDaemon($id_daemon);
+        static::assertSame($id_e, $daemon->id_e);
+    }
+
+    public function testSetDaemonState(): void
+    {
+        $id_daemon = $this->daemonSQL->insertDaemon(1);
+        $daemon = $this->daemonSQL->getDaemon($id_daemon);
+        static::assertSame(Daemon::STATE_INACTIVE, $daemon->state);
+        $this->daemonSQL->setDaemonState($id_daemon, Daemon::STATE_ACTIVE);
+        $daemon = $this->daemonSQL->getDaemon($id_daemon);
+        static::assertSame(Daemon::STATE_ACTIVE, $daemon->state);
+    }
+
+    public function testAllocateWorkers(): void
+    {
+        $id_daemon = $this->daemonSQL->insertDaemon(1);
+        $this->daemonSQL->allocateWorkers($id_daemon, 42);
+        $daemon = $this->daemonSQL->getDaemon($id_daemon);
+        static::assertSame(42, $daemon->nb_workers);
+    }
+
+    public function testRefreshAvailableWorkers(): void
+    {
+        $id_daemon = $this->daemonSQL->insertDaemon(1);
+        $this->daemonSQL->allocateWorkers($id_daemon, 4);
+        $this->daemonSQL->refreshAvailableWorkers();
+        $globalDaemon = $this->daemonSQL->getGlobalDaemon();
+        static::assertSame(1, $globalDaemon->nb_workers);
+    }
+
+    public function testDeleteDaemon(): void
+    {
+        static::assertNotNull($this->daemonSQL->insertDaemon(5));
+        $this->daemonSQL->deleteDaemon(5);
+        static::assertNull($this->daemonSQL->getDaemon(5));
+    }
+
+    public function testGetAllRunningDaemons(): void
+    {
+        $id_daemon = $this->daemonSQL->insertDaemon(1);
+        $this->daemonSQL->setDaemonState($id_daemon, Daemon::STATE_ACTIVE);
+        $id_daemon = $this->daemonSQL->insertDaemon(2);
+        $this->daemonSQL->setDaemonState($id_daemon, Daemon::STATE_INACTIVE);
+        $daemons = $this->daemonSQL->getRunningDaemons();
+        static::assertCount(1, $daemons);
+    }
+
+    public function testGetNbSharedWorkers(): void
+    {
+        static::assertSame((int) NB_WORKERS, $this->daemonSQL->getNbSharedWorkers());
+    }
+
+    public function testGetAllocatedWorkers(): void
+    {
+        $id_daemon = $this->daemonSQL->insertDaemon(1);
+        $this->daemonSQL->allocateWorkers($id_daemon, 10);
+        $id_daemon = $this->daemonSQL->insertDaemon(2);
+        $this->daemonSQL->allocateWorkers($id_daemon, 5);
+        static::assertSame(15, $this->daemonSQL->getNbAllocatedWorkers());
+    }
+}